# 后端OpenAPI草案_v1.yaml
openapi: 3.0.3
info:
  title: ChuanYi Backend API (Draft V1)
  version: 1.0.0-draft
  description: |
    Draft API contract for ChuanYi app.
    Scope includes DIY flow, order flow, and added goods/cart flow.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/order/create:
    post:
      tags: [Order]
      summary: Create order from DIY snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrderCreateData'
  /api/order/create-from-cart:
    post:
      tags: [Order]
      summary: Create order from cart items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateFromCartRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrderCreateData'
  /api/order/create-direct-goods:
    post:
      tags: [Order]
      summary: Create order directly from goods sku
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateDirectGoodsRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrderCreateData'
...
  /api/order/list:
    get:
      tags: [Order]
      summary: Get order list
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - in: query
          name: pageNo
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrderPage'
  /api/order/{orderNo}:
    get:
      tags: [Order]
      summary: Get order detail
      parameters:
        - in: path
          name: orderNo
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrderDetail'
  /api/order/{orderNo}/cancel:
    post:
      tags: [Order]
      summary: Cancel order
      parameters:
        - in: path
          name: orderNo
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderActionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrderActionData'
  /api/order/{orderNo}/ship:
    post:
      tags: [Order]
      summary: Ship order
      parameters:
        - in: path
          name: orderNo
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderShipRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrderShipData'
  /api/order/{orderNo}/confirm-receipt:
    post:
      tags: [Order]
      summary: Confirm receipt
      parameters:
        - in: path
          name: orderNo
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderActionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrderActionData'
...
  /api/admin/order/list:
    get:
      tags: [Admin]
      summary: List orders for admin
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: pageNo
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/admin/order/{orderNo}:
    get:
      tags: [Admin]
      summary: Get admin order detail
      parameters:
        - in: path
          name: orderNo
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/admin/order/{orderNo}/cancel:
    post:
      tags: [Admin]
      summary: Cancel order by admin
      parameters:
        - in: path
          name: orderNo
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                remark:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
  /api/admin/order/{orderNo}/ship:
    post:
      tags: [Admin]
      summary: Ship order by admin
      parameters:
        - in: path
          name: orderNo
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                carrierCode:
                  type: string
                trackingNo:
                  type: string
                remark:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        requestId:
          type: string
          example: 4f6a3d1f8b5545bb
        data:
          nullable: true
...
    PageMeta:
      type: object
      properties:
        pageNo:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64
...
    Address:
      type: object
      properties:
        id:
          type: integer
          format: int64
        receiverName:
          type: string
        receiverPhone:
          type: string
        province:
          type: string
        city:
          type: string
        district:
          type: string
        detailAddress:
          type: string
        isDefault:
          type: boolean
...
    OrderCreateRequest:
      type: object
      required: [snapshotNo, addressId]
      properties:
        snapshotNo:
          type: string
        addressId:
          type: integer
          format: int64
        remark:
          type: string
          maxLength: 45
    OrderCreateFromCartRequest:
      type: object
      required: [addressId]
      properties:
        addressId:
          type: integer
          format: int64
        remark:
          type: string
          maxLength: 45
        cartItemIds:
          type: array
          items:
            type: integer
            format: int64
    OrderCreateDirectGoodsRequest:
      type: object
      required: [skuId, addressId]
      properties:
        skuId:
          type: integer
          format: int64
        qty:
          type: integer
          minimum: 1
          default: 1
        addressId:
          type: integer
          format: int64
        remark:
          type: string
          maxLength: 45
    OrderActionRequest:
      type: object
      properties:
        remark:
          type: string
          maxLength: 128
    OrderShipRequest:
      type: object
      required: [carrierCode, trackingNo]
      properties:
        carrierCode:
          type: string
        trackingNo:
          type: string
        remark:
          type: string
          maxLength: 128
    OrderCreateData:
      type: object
      properties:
        orderNo:
          type: string
        orderStatus:
          $ref: '#/components/schemas/OrderStatus'
        payAmount:
          type: number
          format: double
        preoccupyStatus:
          type: string
          enum: [LOCKED, FAILED]
    ApiResponseOrderCreateData:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OrderCreateData'
    OrderActionData:
      type: object
      properties:
        orderNo:
          type: string
        orderStatus:
          $ref: '#/components/schemas/OrderStatus'
    OrderShipData:
      type: object
      properties:
        orderNo:
          type: string
        orderStatus:
          $ref: '#/components/schemas/OrderStatus'
        carrierCode:
          type: string
        trackingNo:
          type: string
    ApiResponseOrderActionData:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OrderActionData'
    ApiResponseOrderShipData:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OrderShipData'
    OrderStatus:
      type: string
      enum:
        - PENDING_PAY
        - PAID_WAIT_SHIP
        - SHIPPED_WAIT_RECEIVE
        - COMPLETED
        - CANCELLED
        - AFTER_SALE
    OrderItem:
      type: object
      properties:
        itemType:
          type: string
          enum: [DIY_SNAPSHOT, NORMAL_GOODS]
        title:
          type: string
        imageUrl:
          type: string
        unitPrice:
          type: number
          format: double
        qty:
          type: integer
        amount:
          type: number
          format: double
        snapshotNo:
          type: string
          nullable: true
        skuId:
          type: integer
          format: int64
          nullable: true
    OrderSummary:
      type: object
      properties:
        orderNo:
          type: string
        orderStatus:
          $ref: '#/components/schemas/OrderStatus'
        payAmount:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time
        itemPreview:
          type: object
          properties:
            title:
              type: string
            imageUrl:
              type: string
    OrderDetail:
      type: object
      properties:
        orderNo:
          type: string
        orderStatus:
          $ref: '#/components/schemas/OrderStatus'
        payStatus:
          type: string
          enum: [UNPAID, PAID, REFUNDING, REFUNDED]
        totalAmount:
          type: number
          format: double
        payAmount:
          type: number
          format: double
        remark:
          type: string
        addressSnapshot:
          $ref: '#/components/schemas/Address'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        shipping:
          $ref: '#/components/schemas/OrderShipping'
        logs:
          type: array
          items:
            $ref: '#/components/schemas/OrderLog'
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
          nullable: true
    OrderShipping:
      type: object
      properties:
        carrierCode:
          type: string
        trackingNo:
          type: string
        remark:
          type: string
        shippedAt:
          type: string
          format: date-time
    OrderLog:
      type: object
      properties:
        action:
          type: string
        fromStatus:
          type: string
          nullable: true
        toStatus:
          type: string
          nullable: true
        remark:
          type: string
        createdAt:
          type: string
          format: date-time
    ApiResponseOrderPage:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/OrderSummary'
                page:
                  $ref: '#/components/schemas/PageMeta'
    ApiResponseOrderDetail:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OrderDetail'
...